+++ docs/HISTORY.rst
@@ -4,8 +4,8 @@ Changelog
 1.2.7 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Use zope.interface decorator.
+  [gforcada]
 
 1.2.6 (2015-05-31)
 ------------------